removed the folders and moved "data" and "view" into a separate modules
