+++ lib/comp.js
