+++ lib/engine.js
@@ -1,30 +1,48 @@
 
+// require dependencies
+var Module = require('./module');
+var File = require('./static');
+var Socket = require('./socket');
+var Request = require('./request');
+var Session = require('./session');
+var Cache = require('./cache');
+
 exports.clone = clone;
 exports.flat = flat;
 exports.deep = deep;
 exports.path = path;
 
-// clone objects
+/**
+ * Clone object. True prototypal inheritance.
+ *
+ * @public
+ * @param {object} The, to be cloned, object.
+ */
 function clone (object) {
 
     // create an empty function
-    function ClonedObject() {}
+    function O() {}
 
     // set prototype to given object
-    ClonedObject.prototype = object;
+    O.prototype = object;
 
     // create new instance of empty function
-    return new ClonedObject();
+    return new O();
 }
 
-// create flat paths (dot.notation)
+/**
+ * Create a flat object {key1: {key2: "value"}} => {"key1.key2": "value"}
+ *
+ * @public
+ * @param {string} The object, which is flattened.
+ */
 function flat (object) {
     var output = {};
     var value;
     var newKey;
 
     // recusrive handler
-    function step(obj, prev) {
+    function step (obj, prev) {
         for (var key in obj) {
             value = obj[key];
             newKey = prev + key;
@@ -47,7 +65,12 @@ function flat (object) {
     return output;
 }
 
-// nest flat pahts (dot.notation) inside an object
+/**
+ * Unflatten dot-notation keys {"key1.key2": "value"} => {key1: {key2: "value"}}
+ *
+ * @public
+ * @param {string} The object, which is unflattened.
+ */
 function deep (object) {
     var result = {};
     var parentObj = result;
@@ -76,7 +99,14 @@ function deep (object) {
     return result;
 }
 
-// get a nested value from a flat path (dot.notation)
+/**
+ * Get a value from a property "path" (dot.notation).
+ * path('Object.key.key.value'[, {'key': {'key': {'value': 123}}}]);
+ *
+ * @public
+ * @param {string} The path in "dot" notation.
+ * @param {object} The data object, which is used to search the path.
+ */
 function path (key, scope) {
 
     if (!path) {
@@ -95,5 +125,4 @@ function path (key, scope) {
     }
 
     return scope;
-
 }
