+++ lib/fingerprint.js
