+++ lib/session.js
