+++ lib/static.js
@@ -86,3 +86,78 @@ function sendClient (res, removeSid) {
         );
     });
 }
+
+// load client module files (http)
+function moduleFiles (event, request, callback) {
+    var self = this;
+
+    // the module name must be almost alphanumeric
+    if (event.pathname.replace(/[^a-z0-9\/\.\-_@]|\.\.\//gi, '') !== event.pathname) {
+        event.headers = {'content-type': 'text/plain'};
+        return callback(404, 'Incorrect data in module request URL');
+    }
+
+    // get module composition
+    compModules.get(
+
+        // create module name
+        event.path.slice(0, 4).join('_'),
+
+        // request user role
+        event.role,
+
+        // callback
+        function (err, module) {
+
+            // check the roles access to the module
+            if (err || !module) {
+                event.headers = {'content-type': 'text/plain'};
+                err = err ? err.toString() : 'Module not found.';
+                return callback(404, err);
+            }
+
+            // save compressed/compiled script in cache and send it to the client
+            fileModule.set(
+
+                // create absolute file path
+                event.path.slice(4).join('/'),
+
+                // callback
+                function (err, script) {
+
+                    if (err) {
+                        event.headers = {'content-type': 'text/plain'};
+                        return callback(404, errtoString());
+                    }
+
+                    event.headers = script.http;
+                    callback(200, script.data);
+                },
+
+                // pass an array of allowed files
+                module.dependencies,
+
+                // prepend module path to the requested file
+                event.path.slice(0, 4).join('/') + '/'
+            );
+        }
+    );
+}
+
+// get mono client (http)
+function _client (event, request, callback) {
+    var self = this;
+
+    // save compressed/compiled script in cache and send it to the client
+    fileClient.set(event.path[0], function (err, script) {
+
+        if (err) {
+            event.headers = {'content-type': 'text/plain'};
+            callback(500, err.toString());
+            return;
+        }
+
+        event.headers = script.http;
+        callback(200, script.data);
+    });
+}
